requirements.txt
absl-py==0.8.1
amqp==2.6.1
appdirs==1.4.3
asgiref==3.2.10
astor==0.8.0
backports.entry-points-selectable==1.1.0
billiard==3.6.4.0
CacheControl==0.12.6
cached-property==1.5.2
celery==4.4.6
certifi==2019.11.28
cffi==1.14.3
chardet==3.0.4
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.3
contextlib2==0.6.0
cryptography==3.2.1
decorator==4.4.2
distlib==0.3.2
distro==1.4.0
Django==3.1.2
django-cors-headers==3.5.0
django-dramatiq==0.9.1
django-opentracing==1.1.0
djangorestframework==3.12.1
dnspython==1.16.0
dramatiq==1.10.0
eventlet==0.31.1
filelock==3.0.12
future==0.18.2
gast==0.2.2
gevent==21.8.0
google-pasta==0.1.7
greenlet==1.1.1
html5lib==1.0.1
idna==2.8
importlib-metadata==3.4.0
ipaddr==2.2.0
jaeger-client==4.4.0
jsonpath-ng==1.5.2
kombu==4.6.11
lockfile==0.12.2
Markdown==3.1.1
msgpack==0.6.2
numpy==1.20.1
opentracing==2.4.0
opt-einsum==3.1.0
packaging==20.3
pandas==1.2.2
pep517==0.8.2
platformdirs==2.2.0
ply==3.11
progress==1.5
prometheus-client==0.9.0
prompt-toolkit==3.0.20
protobuf==3.10.0
pyasn1==0.4.8
pycparser==2.20
psycopg2==2.9.2
pyparsing==2.4.6
python-dateutil==2.8.1
redis==3.5.3
requests==2.22.0
retrying==1.3.3
six==1.14.0
sqlparse==0.4.1
tensorflow-estimator==2.1.0
termcolor==1.1.0
threadloop==1.0.2
thrift==0.13.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.25.8
vine==1.3.0
virtualenv==20.7.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==0.16.0
wrapt==1.11.2
zipp==3.4.0
zope.event==4.5.0
zope.interface==5.4.0
psutil==5.8.0
setuptools==39.0.1
djangorestframework-jwt==1.11.0
zipfile36==0.1.3